Change saving file system.
However have realised that fit of geometric to quadratic is not optimal i.e. there are gaps on the side. Given that the curve ought to be an ellipse given the geometry of a cone being cut by a plane. This means a version 4 which calibrates a ellipse will be in order